orders-app/requirements.txt
kafka-python==2.0.2
requests==2.28.1